feat!: use generics for entities
When creating entities, you now need to specify a
type (int, float, etc.) that represents the value
for the entity. This is used to ensure that
fields/parameters for the entity are of the
correct type, where applicable. For entities where
the type does not matter, any value can be
specified.