feat(web)!: switch to resty for web requests
BREAKING CHANGES:

This commit switches the code for issuing web
requests to use the resty library. It supports
basic GET and POST requests with a default client
for each request. Apps can also define their own
resty client for customised requests and place it
in a context for use by the request code.